+++ instrumentation/github.com/gorilla/mux/otelmux/config.go
@@ -28,6 +28,7 @@ type config struct {
 	spanNameFormatter func(string, *http.Request) string
 	PublicEndpoint    bool
 	PublicEndpointFn  func(*http.Request) bool
+	Filters           []Filter
 }
 
 // Option specifies instrumentation configuration options.
@@ -40,6 +41,10 @@ type optionFunc func(*config)
 func (o optionFunc) apply(c *config) {
 	o(c)
 }
+
+// Filter is a predicate used to determine whether a given http.request should
+// be traced. A Filter must return true if the request should be traced.
+type Filter func(*http.Request) bool
 
 // WithPublicEndpoint configures the Handler to link the span with an incoming
 // span context. If this option is not provided, then the association is a child
@@ -89,5 +94,17 @@ func WithTracerProvider(provider oteltrace.TracerProvider) Option {
 func WithSpanNameFormatter(fn func(routeName string, r *http.Request) string) Option {
 	return optionFunc(func(cfg *config) {
 		cfg.spanNameFormatter = fn
+	})
+}
+
+// WithFilter adds a filter to the list of filters used by the handler.
+// If any filter indicates to exclude a request then the request will not be
+// traced. All filters must allow a request to be traced for a Span to be created.
+// If no filters are provided then all requests are traced.
+// Filters will be invoked for each processed request, it is advised to make them
+// simple and fast.
+func WithFilter(f Filter) Option {
+	return optionFunc(func(c *config) {
+		c.Filters = append(c.Filters, f)
 	})
 }